requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
Pillow==10.0.0
pydantic==2.3.0
gunicorn==21.2.0
streamlit>=1.28.0
celery==5.3.1
